// src/app/shared/generated/model/contextKind.ts
/**
 * onecx-data-orchestrator-bff
 * OneCX Data Orchestrator BFF
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: tkit_dev@1000kit.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum ContextKind {
    Data = 'Data',
    Database = 'Database',
    KeycloakClient = 'KeycloakClient',
    Microfrontend = 'Microfrontend',
    Microservice = 'Microservice',
    Permission = 'Permission',
    Product = 'Product',
    Slot = 'Slot'
}